Sleep and Kill error checked

+++ myshell.cpp
@@ -37,13 +37,24 @@ int main() {
             }
         }
         else if(strcmp(token, "kill") == 0){
+
             const char* pid = strtok(NULL, d);
             const char* sig = strtok(NULL, d);
-            killProcess(pid, sig);
+            if(pid == NULL || sig == NULL){
+                printf("Error: null argument. Process id and signal number needed. \n");
+            }
+            else {
+                killProcess(pid, sig);
+            }
         }
         else if(strcmp(token, "sleep")==0){
             const char* seconds = strtok(NULL, d);
-            napTime(seconds);
+            if(seconds == NULL){
+                printf("Error: seconds must be declared \n");
+            }
+            else {
+                napTime(seconds);
+            }
         }
         else if (strcmp(token, "cat") == 0) {
             char* filename = strtok(NULL, d);
